refactor: move defineRuntimeConfig to `utils`
(cherry picked from commit cf36fe1d2901029134813e9e2b2cdf882d9e9e6c)
